test: :white_check_mark: Add ARDF tests and restructure file structure
